+++ Libplanet/Store/StoreExtension.cs
@@ -12,40 +12,6 @@ namespace Libplanet.Store
 {
     public static class StoreExtension
     {
-        #pragma warning disable MEN002
-        /// <summary>
-        /// Looks up a state reference, which is a block's <see cref="Block{T}.Hash"/> that contains
-        /// an action mutating the <paramref name="address"/>' state.
-        /// </summary>
-        /// <param name="store">The store object expected to contain the state reference.</param>
-        /// <param name="chainId">The chain ID to look up a state reference.</param>
-        /// <param name="address">The <see cref="Address"/> to look up.</param>
-        /// <param name="lookupUntil">The upper bound (i.e., the latest block) of the search range.
-        /// <see cref="Block{T}"/>s after <paramref name="lookupUntil"/> are ignored.</param>
-        /// <returns>Returns a nullable tuple consisting of <see cref="Block{T}.Hash"/> and
-        /// <see cref="Block{T}.Index"/> of the <see cref="Block{T}"/> with the state of the
-        /// address.</returns>
-        /// <typeparam name="T">An <see cref="IAction"/> class used with
-        /// <paramref name="lookupUntil"/>.</typeparam>
-        /// <seealso cref="IStore.StoreStateReference(Guid, IImmutableSet{Address}, HashDigest{SHA256}, long)"/>
-        /// <seealso cref="IStore.IterateStateReferences(Guid, Address, long?, long?, int?)"/>
-        #pragma warning restore MEN002
-        public static Tuple<HashDigest<SHA256>, long> LookupStateReference<T>(
-            this IStore store,
-            Guid chainId,
-            Address address,
-            Block<T> lookupUntil)
-            where T : IAction, new()
-        {
-            if (lookupUntil is null)
-            {
-                throw new ArgumentNullException(nameof(lookupUntil));
-            }
-
-            return store.IterateStateReferences(chainId, address, lookupUntil.Index, limit: 1)
-                    .FirstOrDefault();
-        }
-
         /// <summary>
         /// Makes a store, <paramref name="to"/>, logically (but not necessarily physically)
         /// identical to another store, <paramref name="from"/>.  As this copies the contents
@@ -70,7 +36,7 @@ namespace Libplanet.Store
                 {
                     Block<NullAction> block = from.GetBlock<NullAction>(blockHash);
                     to.PutBlock(block);
-                    IImmutableDictionary<Address, IValue> states = from.GetBlockStates(blockHash);
+                    IImmutableDictionary<string, IValue> states = from.GetBlockStates(blockHash);
                     to.SetBlockStates(blockHash, states);
                     to.AppendIndex(chainId, blockHash);
                 }
@@ -80,14 +46,14 @@ namespace Libplanet.Store
                     to.IncreaseTxNonce(chainId, kv.Key, kv.Value);
                 }
 
-                foreach (Address address in from.ListAddresses(chainId))
+                foreach (string key in from.ListStateKeys(chainId))
                 {
-                    foreach (var pair in from.IterateStateReferences(chainId, address).Reverse())
+                    foreach (var pair in from.IterateStateReferences(chainId, key).Reverse())
                     {
                         pair.Deconstruct(out HashDigest<SHA256> refHash, out long refIndex);
                         to.StoreStateReference(
                             chainId,
-                            ImmutableHashSet.Create(address),
+                            ImmutableHashSet.Create(key),
                             refHash,
                             refIndex
                         );